Refactor IdlingResource
