--- RxJavaTemplate/app/src/production/java/co/nimblehq/rxjava/IdlingResource.kt
@@ -1,6 +1,0 @@
-package co.nimblehq.rxjava
-
-object IdlingResource {
-
-    fun increaseOrDecrease(increase: Boolean) {}
-}
